others/joi2008yo/e/src/test/java/com/solver/atcoder/others/joi2008yo/e/MainTest.java
package com.solver.atcoder.others.joi2008yo.e;

import java.io.ByteArrayInputStream;
import java.io.ByteArrayOutputStream;
import java.io.PrintStream;

import org.junit.Test;

import static org.hamcrest.CoreMatchers.is;
import static org.hamcrest.MatcherAssert.assertThat;

public class MainTest {

  @Test
  public void test() {
    String input = "2 5\n"
        + "0 1 0 1 0\n"
        + "1 0 0 0 1";
    String expected = "9";
    
    assertThat(execute(input + "\n"), is(expected + "\n"));
  }

  @Test
  public void test1() {
    String input = "3 6\n"
        + "1 0 0 0 1 0\n"
        + "1 1 1 0 1 0\n"
        + "1 0 1 1 0 1";
    String expected = "15";

    assertThat(execute(input + "\n"), is(expected + "\n"));
  }

  private String execute(String input) {
    ByteArrayOutputStream os = new ByteArrayOutputStream();
    Main.solve(
        new ByteArrayInputStream(input.getBytes()), new PrintStream(os)
    );
    return os.toString().replace(System.lineSeparator(), "\n");
  }

}